Добавих метод за добавяне на оценка и за гет на всички ученици по предмет (оценката вече е decimal)

+++ Services/DAO/MarkDAO.cs
@@ -8,7 +8,7 @@ namespace Services.DAO
 {
     public partial class MarkDAO
     {
-        public int AddMark(int markValue, int subjectId, int studentId, string comment)
+        public int AddMark(decimal markValue, int subjectId, int studentId, string comment)
         {
             if (markValue < 2 || markValue > 6)
             {
